fix: correct the diff heat map and stop the geometry matcher inventing absences
The heat map multiplied a uint8 red channel by 3 before widening, so it wrapped
modulo 256: a difference of 86 rendered as 2 and 200 as 88, meaning the most
different regions appeared darkest in the images a reviewer is told to inspect.
Verified numerically against the wrapping values.

The geometry matcher reported elements as missing that were present. The audit
dumps a quick-panel tab's visible text in caps and its container's content-desc
in title case, so every tab produced a phantom absence; authorized D014 renames
(TWITCH.TV -> PLATFORM A, Streamlabs -> Dashboard A and so on) were reported as
missing for the same reason; and a label carried on both a TextView and its
parent was counted twice. Keys are now case-folded, renames pair through an
alias table, same-bounds duplicates index once, and about-box copy whose length
differs by authorization is excluded rather than counted. Matched elements rise
1041 -> 1060, unmatched fall 238 -> 217, and states with nothing unmatched go
77 -> 81.

Scoring more elements lowers the within-2px percentage, because 19 newly
visible elements joined the denominator, so the ratchet baseline is re-recorded
with that reason and the coverage numbers that justify it. -UpdateBaseline now
refuses to lower the bar unless a reason is given.

Also: build_final_report exits with a one-line explanation instead of a
TypeError when no results exist, and the anchor extractor logs the surfaces it
skips and admits a straddling first row for selection dialogs.

+++ replica-app/scripts/visual_compare.py
@@ -116,7 +116,10 @@ def main() -> int:
     Image.blend(reference_image, aligned_image, 0.5).save(args.overlay)
     heat = np.max(absolute, axis=2)
     heat_rgb = np.zeros_like(reference)
-    heat_rgb[..., 0] = np.clip(heat * 3, 0, 255)
+    # heat is uint8: multiplying before widening wraps modulo 256, so a diff of
+    # 86 rendered as red 2 and the *worst* regions came out near-black - in the
+    # exact images the report tells a reviewer to inspect.
+    heat_rgb[..., 0] = np.clip(heat.astype(np.int16) * 3, 0, 255)
     heat_rgb[..., 1] = np.clip(np.maximum(heat.astype(np.int16) - 85, 0) * 3, 0, 255)
     heat_rgb[~valid] = np.array([32, 32, 32], dtype=np.uint8)
     args.diff.parent.mkdir(parents=True, exist_ok=True)
